// packages/frontend/src/api/financialModelingPrep/types.ts
export type IncomeStatement = {
   date: string
   symbol: string
   reportedCurrency: string
   fillingDate: string
   acceptedDate: string
   period: string
   calendarYear?: string
   revenue: number
   costOfRevenue: number
   grossProfit: number
   grossProfitRatio: number
   researchAndDevelopmentExpenses: number
   generalAndAdministrativeExpenses: number
   sellingAndMarketingExpenses: number
   otherExpenses: number
   operatingExpenses: number
   costAndExpenses: number
   interestExpense: number
   depreciationAndAmortization: number
   ebitda: number
   ebitdaratio: number
   operatingIncome: number
   operatingIncomeRatio: number
   totalOtherIncomeExpensesNet: number
   incomeBeforeTax: number
   incomeBeforeTaxRatio: number
   incomeTaxExpense: number
   netIncome: number
   netIncomeRatio: number
   eps: number
   epsdiluted: number
   weightedAverageShsOut: number
   weightedAverageShsOutDil: number
   link: string
   finalLink: string
}